requirements.txt
streamlit
feedparser
pandas
langchain-google-genai
python-dotenv
